backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
openai==1.30.1
azure-search-documents==11.6.0b4
azure-identity==1.16.0
azure-storage-blob==12.20.0
azure-cosmos==4.7.0
langchain==0.2.1
langchain-openai==0.1.8
langgraph==0.0.55
llama-index==0.10.42
sentence-transformers==3.0.0
redis[asyncio]==5.0.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-docx==1.1.0
PyMuPDF==1.24.3
openpyxl==3.1.2
structlog==24.1.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
